# internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Nexodus API Server.",
        "title": "Nexodus API",
        "contact": {
            "name": "The Nexodus Authors",
            "url": "https://github.com/nexodus-io/nexodus/issues"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/devices": {
            "get": {
                "description": "Lists all devices",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Devices"
                ],
                "summary": "List Devices",
                "operationId": "ListDevices",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Device"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.BaseError"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/models.BaseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.InternalServerError"
                        }
                    }
                }
            },
            "post": {
                "description": "Adds a new device",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Devices"
                ],
                "summary": "Add Devices",
                "operationId": "CreateDevice",
                "parameters": [
                    {
                        "description": "Add Device",
                        "name": "Device",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AddDevice"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Device"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.BaseError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.BaseError"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/models.ConflictsError"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/models.BaseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.InternalServerError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.AddDevice": {
            "type": "object",
            "properties": {
                "child_prefix": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "172.16.42.0/24"
                    ]
                },
                "discovery": {
                    "type": "boolean"
                },
                "endpoint_local_address_ip4": {
                    "type": "string",
                    "example": "1.2.3.4"
                },
                "endpoints": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Endpoint"
                    }
                },
                "hostname": {
                    "type": "string",
                    "example": "myhost"
                },
                "os": {
                    "type": "string"
                },
                "public_key": {
                    "type": "string"
                },
                "relay": {
                    "type": "boolean"
                },
                "security_group_id": {
                    "type": "string"
                },
                "symmetric_nat": {
                    "type": "boolean"
                },
                "tunnel_ip": {
                    "type": "string",
                    "example": "1.2.3.4"
                },
                "tunnel_ip_v6": {
                    "type": "string",
                    "example": "200::1"
                },
                "vpc_id": {
                    "type": "string",
                    "example": "694aa002-5d19-495e-980b-3d8fd508ea10"
                }
            }
        },
...
        "models.BaseError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "something bad"
                }
            }
        },
        "models.ConflictsError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "something bad"
                },
                "id": {
                    "type": "string",
                    "example": "a1fae5de-dd96-4b20-8362-95f6a574c4b1"
                }
            }
        },
        "models.Device": {
            "type": "object",
            "properties": {
                "allowed_ips": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "bearer_token": {
                    "description": "the token nexd should use to reconcile device state.",
                    "type": "string"
                },
                "child_prefix": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "discovery": {
                    "type": "boolean"
                },
                "endpoint_local_address_ip4": {
                    "type": "string"
                },
                "endpoints": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Endpoint"
                    }
                },
                "hostname": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": "aa22666c-0f57-45cb-a449-16efecc04f2e"
                },
                "online": {
                    "type": "boolean"
                },
                "online_at": {
                    "type": "string"
                },
                "organization_prefix": {
                    "type": "string"
                },
                "organization_prefix_v6": {
                    "type": "string"
                },
                "os": {
                    "type": "string"
                },
                "owner_id": {
                    "type": "string"
                },
                "public_key": {
                    "type": "string"
                },
                "relay": {
                    "type": "boolean"
                },
                "revision": {
                    "type": "integer"
                },
                "security_group_id": {
                    "type": "string"
                },
                "symmetric_nat": {
                    "type": "boolean"
                },
                "tunnel_ip": {
                    "type": "string"
                },
                "tunnel_ip_v6": {
                    "type": "string"
                },
                "vpc_id": {
                    "type": "string",
                    "example": "694aa002-5d19-495e-980b-3d8fd508ea10"
                }
            }
        },
...
        "models.Endpoint": {
            "type": "object",
            "properties": {
                "address": {
                    "description": "IP address and port of the endpoint.",
                    "type": "string",
                    "example": "10.1.1.1:51820"
                },
                "distance": {
                    "description": "Distance in milliseconds from the node to the ip address",
                    "type": "integer",
                    "example": 12
                },
                "source": {
                    "description": "How the endpoint was discovered",
                    "type": "string"
                }
            }
        },
        "models.InternalServerError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "something bad"
                },
                "trace_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "OAuth2Implicit": {
            "type": "oauth2",
            "flow": "implicit",
            "authorizationUrl": "https://auth.try.nexodus.127.0.0.1.nip.io/",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "user": " Grants read and write access to resources owned by this user"
            }
        }
    }
}